+++ desktop/src/main/kotlin/vn/loi/learning/desktop/ui/designsystem/components/base/LESurface.kt
@@ -6,12 +6,14 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.Dp
 import vn.loi.learning.desktop.ui.theme.LETheme
 
 @Composable
 fun LESurface(
     variant: LESurfaceVariant,
     modifier: Modifier = Modifier,
+    contentPadding: Dp? = null,
     content: @Composable ColumnScope.() -> Unit
 ) {
     val style = resolveSurfaceStyle(LETheme.colors, variant)
@@ -24,7 +26,7 @@ fun LESurface(
         shadowElevation = LETheme.elevation.elevation1
     ) {
         Column(
-            modifier = Modifier.padding(LETheme.spacing.space5),
+            modifier = Modifier.padding(contentPadding ?: LETheme.spacing.space5),
             content = content
         )
     }